requirements.txt
python-dotenv==0.20.0
requests==2.31.0
discord.py==2.3.1
